doh. make u32 actually 32 bits wide (instead of 64)

+++ jtag-driver.h
@@ -19,7 +19,7 @@
 
 typedef unsigned char u8;
 typedef unsigned short u16;
-typedef unsigned long u32;
+typedef unsigned int u32;
 
 typedef struct JDRV JDRV;
